requirements.txt
# requirements.txt
anyio==4.6.2.post1
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.100.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.1.0
numpy==2.1.3
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.27.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.23.1
watchfiles==0.24.0
websockets==14.1
gunicorn==21.2.0 #ケントさん
